LambdaCDM is fitted with CC+DESIBAO+PLanck data
Here, we fitted the same model with same data set using two different sampler. The polychord and emcee results are very similar. Since we have used uniform priors in radiation density in emcee, hence, we are getting non-gaussian posterior distribution which can be easily marginalized. However, in case of polychord, we are using a Gaussian prior on Omega_r, therefore, the posterior distribution looks incredibly good.